Extract TranslationStatusBanner.vue
