+++ src/utils/local.js
@@ -4,25 +4,11 @@ const localStorageExists =
   typeof window !== 'undefined' && 'localStorage' in window
 
 const local = {
-  /**
-   * Gets the value from localStorage if window is defined,
-   * or falls back on defaultValue.
-   * @param {string} key
-   * @param {string|null} defaultValue - value used if localStorage is not available,
-   * or if the key is not in localStorage
-   * @returns {string|null}
-   */
-  get(key, defaultValue = null) {
-    return localStorageExists ? localStorage.getItem(key) : defaultValue
+  get(key) {
+    return localStorageExists ? localStorage.getItem(key) : null
   },
-  /**
-   * Sets the value for key in localStorage.
-   * @param {string} key
-   * @param {string} value
-   * @returns {void|null} - null is returned if localStorage is not available.
-   */
   set(key, value) {
-    localStorageExists ? localStorage.setItem(key, value) : null
+    return localStorageExists ? localStorage.setItem(key, value) : null
   },
 }
